fix(web): tighten bottle tables and tab dividers
